+++ lib/generators/cms/kickstart/kickstart_generator.rb
@@ -254,19 +254,25 @@ module Cms
         Rails::Generators.invoke('cms:component:search')
         Rails::Generators.invoke('cms:component:login_page')
         Rails::Generators.invoke('cms:component:sitemap')
-        Rails::Generators.invoke('cms:widget:text')
-        Rails::Generators.invoke('cms:widget:image')
+
+        unless examples?
+          Rails::Generators.invoke('cms:widget:text')
+          Rails::Generators.invoke('cms:widget:image')
+          Rails::Generators.invoke('cms:widget:headline')
+        end
       end
 
       def create_example_content
         if examples?
+          Rails::Generators.invoke('cms:widget:teaser', ['--example'])
+          Rails::Generators.invoke('cms:widget:image', ['--example'])
+          Rails::Generators.invoke('cms:widget:headline', ['--example'])
+          Rails::Generators.invoke('cms:widget:maps', ['--example'])
+          Rails::Generators.invoke('cms:widget:text', ['--example'])
+
           Rails::Generators.invoke('cms:component:profile_page', ['--cms_path=/website/en'])
           Rails::Generators.invoke('cms:component:contact_page', ['--cms_path=/website/en'])
           Rails::Generators.invoke('cms:component:blog', ['--cms_path=/website/en'])
-          Rails::Generators.invoke('cms:widget:teaser')
-          Rails::Generators.invoke('cms:widget:maps')
-
-          migration_template('create_examples.rb', 'cms/migrate/create_examples.rb')
         end
       end