+++ lib/generators/cms/widget/maps/maps_generator.rb
@@ -2,14 +2,16 @@ module Cms
   module Generators
     module Widget
       class MapsGenerator < ::Rails::Generators::Base
+        include Example
+
         source_root File.expand_path('../templates', __FILE__)
 
         SUPPORTED_PROVIDER = %w(google_maps)
 
         class_option :provider,
-          :type => :string,
-          :default => 'google_maps',
-          :desc => "Select what maps provider to use. (#{SUPPORTED_PROVIDER.join(' | ')})"
+          type: :string,
+          default: 'google_maps',
+          desc: "Select what maps provider to use. (#{SUPPORTED_PROVIDER.join(' | ')})"
 
         def validate_provider
           unless SUPPORTED_PROVIDER.include?(options[:provider])
@@ -23,9 +25,8 @@ module Cms
         end
 
         def run_generator_for_selected_provider
-          Rails::Generators.invoke(
-            "cms:widget:maps:#{options[:provider]}"
-          )
+          provider_options = Thor::Options.to_switches(options).split
+          Rails::Generators.invoke("cms:widget:maps:#{options[:provider]}", provider_options)
         end
       end
     end