+++ lib/generators/cms/widget/slideshare/slideshare_generator.rb
@@ -2,25 +2,18 @@ module Cms
   module Generators
     module Widget
       class SlideshareGenerator < ::Rails::Generators::Base
+        include Example
+        include Migration
+
         source_root File.expand_path('../templates', __FILE__)
 
         def create_migration
           begin
             Widget::ApiGenerator.new(behavior: behavior) do |widget|
-              widget.name = 'SlideshareWidget'
+              widget.name = obj_class_name
               widget.icon = '&#xF01A;'
               widget.description = 'Creates a widget that shows a slide from slideshare.'
               widget.attributes = [
-                {
-                  name: 'headline',
-                  type: :string,
-                  title: 'Headline',
-                },
-                {
-                  name: 'content',
-                  type: :html,
-                  title: 'content',
-                },
                 {
                   name: 'source',
                   type: :linklist,
@@ -35,10 +28,20 @@ module Cms
           end
         end
 
+        def create_example
+          example_migration_template(obj_class_name.underscore)
+        end
+
         def notice
           if behavior == :invoke
             log(:migration, 'Make sure to run "rake cms:migrate" to apply CMS changes')
           end
+        end
+
+        private
+
+        def obj_class_name
+          'SlideshareWidget'
         end
       end
     end