+++ lib/generators/cms/widget/text/text_generator.rb
@@ -2,20 +2,18 @@ module Cms
   module Generators
     module Widget
       class TextGenerator < ::Rails::Generators::Base
+        include Example
+        include Migration
+
         source_root File.expand_path('../templates', __FILE__)
 
         def create_widget
           begin
             Widget::ApiGenerator.new(behavior: behavior) do |widget|
-              widget.name = 'TextWidget'
+              widget.name = obj_class_name
               widget.icon = '&#xF058;'
-              widget.description = 'Creates a simple widget with headline and content.'
+              widget.description = 'Creates a simple widget with content.'
               widget.attributes = [
-                {
-                  name: 'headline',
-                  type: :string,
-                  title: 'Headline',
-                },
                 {
                   name: 'content',
                   type: :html,
@@ -29,10 +27,20 @@ module Cms
           end
         end
 
+        def create_example
+          example_migration_template(obj_class_name.underscore)
+        end
+
         def notice
           if behavior == :invoke
             log(:migration, 'Make sure to run "rake cms:migrate" to apply CMS changes')
           end
+        end
+
+        private
+
+        def obj_class_name
+          'TextWidget'
         end
       end
     end