+++ lib/generators/cms/widget/video/templates/app/models/video_widget.rb
@@ -1,8 +1,6 @@
 class VideoWidget < Obj
   include Widget
 
-  cms_attribute :headline, type: :string
-  cms_attribute :content, type: :html
   cms_attribute :source, type: :linklist, max_size: 1
   cms_attribute :width, type: :integer
   cms_attribute :height, type: :integer