+++ lib/generators/cms/widget/video/video_generator.rb
@@ -2,6 +2,9 @@ module Cms
   module Generators
     module Widget
       class VideoGenerator < ::Rails::Generators::Base
+        include Example
+        include Migration
+
         source_root File.expand_path('../templates', __FILE__)
 
         def video_tools
@@ -44,20 +47,10 @@ module Cms
         def create_migration
           begin
             Widget::ApiGenerator.new(behavior: behavior) do |widget|
-              widget.name = 'VideoWidget'
+              widget.name = obj_class_name
               widget.icon = '&#xF062;'
               widget.description = 'Displays a video player for the given video file.'
               widget.attributes = [
-                {
-                  name: 'headline',
-                  type: :string,
-                  title: 'Headline',
-                },
-                {
-                  name: 'content',
-                  type: :html,
-                  title: 'Content',
-                },
                 {
                   name: 'source',
                   type: :linklist,
@@ -96,10 +89,20 @@ module Cms
           end
         end
 
+        def create_example
+          example_migration_template(obj_class_name.underscore)
+        end
+
         def notice
           if behavior == :invoke
             log(:migration, 'Make sure to run "rake cms:migrate" to apply CMS changes')
           end
+        end
+
+        private
+
+        def obj_class_name
+          'VideoWidget'
         end
       end
     end