+++ spec/generators/widget/login_generator_spec.rb
@@ -8,6 +8,7 @@ describe Cms::Generators::Widget::LoginGenerator do
   include GeneratorSpec::TestCase
 
   destination File.expand_path('../../../../tmp/generators', __FILE__)
+  arguments ['--example']
 
   before do
     prepare_destination
@@ -47,14 +48,13 @@ describe Cms::Generators::Widget::LoginGenerator do
 
             directory 'migrate' do
               migration 'create_login_widget'
+              migration 'create_login_widget_example'
             end
           end
         end
 
         directory 'models' do
           file 'login_widget.rb' do
-            contains 'cms_attribute :headline, type: :string'
-            contains 'cms_attribute :content, type: :html'
             contains 'include Widget'
           end
         end