+++ spec/generators/widget/slider_generator_spec.rb
@@ -8,6 +8,7 @@ describe Cms::Generators::Widget::SliderGenerator do
   include GeneratorSpec::TestCase
 
   destination File.expand_path('../../../../tmp/generators', __FILE__)
+  arguments ['--example']
 
   before do
     prepare_destination
@@ -50,6 +51,7 @@ describe Cms::Generators::Widget::SliderGenerator do
 
             directory 'migrate' do
               migration 'create_slider_widget'
+              migration 'create_slider_widget_example'
             end
           end
         end