+++ spec/generators/widget/slideshare_generator_spec.rb
@@ -8,6 +8,7 @@ describe Cms::Generators::Widget::SlideshareGenerator do
   include GeneratorSpec::TestCase
 
   destination File.expand_path('../../../../tmp/generators', __FILE__)
+  arguments ['--example']
 
   before do
     prepare_destination
@@ -34,14 +35,13 @@ describe Cms::Generators::Widget::SlideshareGenerator do
 
             directory 'migrate' do
               migration 'create_slideshare_widget'
+              migration 'create_slideshare_widget_example'
             end
           end
         end
 
         directory 'models' do
           file 'slideshare_widget.rb' do
-            contains 'cms_attribute :content, type: :html'
-            contains 'cms_attribute :headline, type: :string'
             contains 'cms_attribute :source, type: :linklist'
             contains 'include Widget'
           end